added db connection
